refactor: page/component/* is only used in that page or move it to global component
