+++ lib/views/components/content/post_votes.dart
