+++ lib/views/components/content/publication_header.dart
@@ -1,16 +1,16 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/services/conversion/human_time_service.dart";
-import "package:proxima/views/components/user_avatar/user_avatar.dart";
+import "package:proxima/views/components/content/user_avatar/user_avatar.dart";
 
 /// This widget is used to display the info bar in the post card.
 /// It contains the user's profile picture and username
 /// and the publication time of the post.
-class PostHeader extends ConsumerWidget {
+class PublicationHeader extends ConsumerWidget {
   static const displayNameTextKey = Key("displayNameText");
   static const publicationDateTextKey = Key("publicationTimeTextKey");
 
-  const PostHeader({
+  const PublicationHeader({
     super.key,
     required this.posterUsername,
     required this.publicationDate,
